initial setup to develop using react, typescript & webpack
Additionaly create an app named `frontend` that will contains the code
for the different react components of the project and the transpiled
javascript code.

As an example, this commit creates a feedback modal using react. The
main component is called <ReportButton /> that renders a button and
a modal to allow users report bugs and make suggestions to admins. We
need to create a folder inside `media` called `screenshot` to put all
images uploaded by users in their feedbacks.

Two things we need to look at closely:
- Enable translation for strings in react
- Think about testing in react (we can use jest + enzyme for this)

This commit fixes (using css) the navbar movement when a modal was
opened.