[command] create dispose command to add domains to disposable.json file
This commit adds a new command named `dispose` to ease the process of
adding a domain to disposable.json file. To use dispose type the
following command in the console:

`$ python manage.py dispose --domain <domain>`

That will add `<domain>` to disposable.json file and sort the domains.

This commit also adds the following domains to disposable.json:
- `maxwinslots24.xyz`
- `onlinecasinogames777.us`
- `taidar.ru`