[palantir] add setting to enable/disable requests logging
Everytime we log, we do it using two really bad things: creating a
thread to handle the logging and storing the request log in the
database :(

During tests, those threads can cause some problems in teardown when the
test class tries to drop the testing database. This is due to the extra
connections to postgres hold by Palantir in a HTTP request (using
`self.client` for instance).

Adding this setting, helps us also to disable the logging if we found
this approach unsuitable in prod.