[permissions] Add table to store specific permissions per user per contest
This PR adds a new table called ContestPermission that stores info about
which user can do in each contest (specifing a certain role).

The table will looks like:

```
+----+------+---------+-------+---------+---------------------+
| pk | user | contest | role  | granted | date                |
+----+------+---------+-------+---------+---------------------+
| 1  | 1    | 5       | judge | 1       | 2020-01-01 00:00:00 |
+----+------+---------+-------+---------+---------------------+
| 2  | 2    | 5       | judge | 1       | 2020-01-02 00:00:00 |
+----+------+---------+-------+---------+---------------------+
| 3  | 1    | 5       | judge | 0       | 2020-01-03 00:00:00 |
+----+------+---------+-------+---------+---------------------+
```

The way this table will be used is that given a triple like
(user, contest, role) we look for the latest entry in the table and
check the granted column. In the sample above, user 2 has `judge` role
on contest 5. However, user 1 lost their `judge` role after being
granted.

This table works as a log and we should only add new rows to it.