Transfer sample cases to the database
The main advantages of this are:

- We don't need to copy the problem's folder to every webserver
- The sample test cases can be edited/removed/added through the web
  application (previously we would have had to copy them to every
  webserver)
- The problem page should load faster since the sample cases come
  already with the problem model, i.e. no need to go find them on the
  hard drive.

The changes are transparent to the final user since the handling of the
sample cases in the "Edit Test Cases" view looks the same, i.e. samples
appear as files

A command `fix_samples` is included (should be called in this way:
`python manage.py fix_samples`) that loads the sample cases previously
stored on the hard drive into the database. Functions related to the
handling of samples were moved to a new file `samples.py`

Original Merge Request !34