[permissions] Add permission functions to ask for roles in contest
This commit adds helper functions that query the ContestPermission table
to check if users has specific roles there. The idea is that we shouldnt
use the `user_is_judge`/`user_is_observer` functions that are user
based only.

This commit also do the following things:
- fix judge can see their hidden submissions in the submission list
- fix query for contests list to call the new gating function (judges)
- added admin view to show permissions filtered by: judges, observers
and active (granted or not).
- remove legacy code (commit from otero squashed here)
- modified existing tests and interacted with the local server