Adding redirect to data server to edit testcases
We are facing the problem that the **Testcases** can not be **uploaded/updated** from the web application due to the fact that it is not hosted on the same server, where the data for the problems is stored (and we don't have the possibility to get some centralized filesystem or storage space at the moment).

This is a simple solution that redirects the user (should be admin or judge) to the replica of the web application hosted in the data server.

**Notes:**
This replica should be used only for this since we do not want to increase the load of the grader server. Moreover, since this replica is located in Germany and is pointing to the DB (only master) that is in New York, it is much slower than the other webservers.
Current grader server url: [http://213.136.83.17](http://213.136.83.17)

Also included in this PR **permissions** for the **role judge** to edit the test cases.

![Un_tren_con_dos_destinos___MatCom_Online_Grader_-_Google_Chrome_2019-09-07_14-15-50](/uploads/7429f397b213b25a233d99a6734c0f8e/Un_tren_con_dos_destinos___MatCom_Online_Grader_-_Google_Chrome_2019-09-07_14-15-50.mp4)

For this to happen we need to put in the ```settings.ini``` file the following section:

```
[redirect]
DATA_SERVER_URL: http://213.136.83.17
```

**Remark:** The replica hosted on the data server serves as a backup plan in case something fails. This is a list of possible bad scenarios during an important Contest that could be solved using this server:

1.  The route (in Cuba) to the current webservers fails: this is a remote case (baned ip, DNS ignoring this route,... whatever crazy reasons). The solution is to give the contestants the URL for the grader server. The grader server is pointing to the production database so the state of the contest won't be lost and participants will be able to continue sending their submissions. Note: this solution doesn't require any technical change.

2.  The web servers fail, or the REDIS server for authentication fails, or any other failure but the database cluster is still alive an accessible: The solution is the same as before.