set default image as fallback if avatar url is not found
This might happen if the user uploads an image and then hard refresh the
page. That second refresh invalidate the cache and the next request can
go to a server where the avatar image doesn't exist. In that case we
want at least show the default user image instead of displaying
something broken.

This commit adds that default image on the onerror event.

Tested in localhost:
-