Fix bug in problem edit and create view
The field for the JSON with the multiple time and memory limits was
not working properly in the problem create and edit views. It was
working ok in the admin panel though.

Validation for the field multiple_limits was included and mog.js
was modified to prevent the corresponding textarea from being
rendered with the markdown editor.