Import account from baylor's .tab files
Ussage:
python manage.py import_baylor \
    --path C:\Users\otero\Downloads\CLC-2018 \
    --prefix 2018cLc \
    --contest 6240

Additionally the option --remove can be used for removing the teams.
This can be used in the testing phase, for example, if we create old
accounts in the production server and the would like to remove them.

The path indicates where the .tab files (School.tab, Site.tab, Team.tab,
Person.tab, TeamPerson.tab) are.

This will load the teams and create the account if they do not exist and
will register them for the specified contest.

For the Local Contests the usage will be the following:
- Call the command with the id of the practice contest
- Call the command with the id of the real contest

After this, the teams will be registered for both contests. After the
second time, a .txt file per site is generated with the password and
accounts of the teams (after the first time too but the second time
overwrites the passwords).