[block] Add BLOCK_PUBLIC_ACTIONS flag and decorator
To block public actions this flag should be set in the settings.ini file:

```
[flags]
BLOCK_PUBLIC_ACTIONS: true
```

The decorator @public_actions_required should be used for any function that should be blocked. This will check if the flag is active and the user is not admin. In this case, will not call the function and show this message:

`This action is currently blocked!`

[Follow Up]: When this flag is active do not allow the user to submit or see any code outside from a running contest (this maybe should be part of this PR but wanted to split it to keep it simple)