Fix filter by pending
Before this, the status Frozen was considering as Pending when filtering. This doesn't make much sense since the frozen submissions are always the last ones,i.e. they always appear on the top of the list.

When an admin or judge in the contest (in case of viewing the submissions from the contest suite) filters by result will see exactly all submissions with that result.

When a judge filters from the global page.... he wont be able to see the submissions that are frozen in the result (the admin can see them). This is because in the global submissions page there could be submissions from a different contest and the query would be much complicated.

A normal user only sees not-frozen submissions unless they belong to him.