Redirect to team submissions on click on team
This PR defines the behaviour "click on team"

Previously: we only had normal teams (with at least one user or more) In this case one can click on any of the users to see for example the submissions of this user.

New case: teams with only one user where only the name is read. In this case one can not click anything. Therefore, here we define that when the user clicks in such a team,... he will be redirected to a page that shows the submissions of the underlying user in the current contest.