Prevent rejudge when submission is currently being graded
Prevent admin or judge to rejudge submission is the current result is
Pending, Compiling or Running.

If there is some problem with the grader and the submission stays as
Compiling or Running it won't be possible to rejudge it. A superadmin
should go to django/admin and set manually the result to Pending. This
is an unlikely case.

This will avoid a silent fail when rejudging a running solution (the
result used to be internal error or runtime error)