Grader catch exceptions and change way of taking submissions to grade
Lock to take the first available pending submission and change its status to Compiling. This is an atomic transaction and the submission is blocked for other graders.

Fixed behaviour when rejudging a submission that is being judged by another grader

Catch database-related exceptions:

The connection to the database was interrupted or could not be established
Race condition in a trigger in the database (TODO: Fix this race condition)