Move create problem link inside a contest
The create button makes more sense inside a contest. This commit do that
and enforce the new problem to be associated with the right contest. The
user will be able to change the contest using the "Edit problem" view.

This commit also checks that the problem is moved only to allowed
contests: the ones the user can add problems to. A user can create a new
problem in a contest if their are either and admin OR a judge of the
specified contest.

The URL to create the problem also changes to the following format:
`/contest/<contest-id>/create-problem`. With this, we will be able to
identify which contest must be related to the newly created problem,
check permissions and redirect properly.