[cleanup] Remove unused slack app
This app was integrated with Slack to report stats, however we are not
using it anymore and it was a security risk since we didn't check for
permissions on incoming requests.

In the future, we could add a more secure way to generate stats
deliverable to any end (e.g emails, slack, discord, messages, etc).