Grader refactorization
This PR will separate the internal logic of the grader (currently using runexe) by implementing a class that exposes the API required from the platform. Ultimately the plan is to replace the runexe executor with another executor that runs on linux so the platform can be moved to linux (since this is the only blocking feature).

Changes introduced by this PR (and how to review it):

The most important change happened at the function `grade_submission(submission, number_of_executions)` on `api/management/commands/grader.py`.

Previously the full logic of the grader (regarding the execution of the submission happened at this point). This is still the case right, however, this logic was wrapped behind a `BaseGrader` object.

One notable change before going deeply inside how `BaseGrader` works, is that compilation previously happened before calling to `grade_submission`, but it makes sense to delegate this process to the grader itself (as it might vary between different implementations). Right now the `grade_submission` function first compile the submission, and then execute it: all submissions labels `[Pending, Compiling, Running, Final-Verdict]` were adjusted properly.

About the new structure:

We created a new folder `api/grader` that contains all the logic regarding the grader, with its multiple utils functions that were moved from several files located a: `api/management/commands/` such as `__checker_backends` and `__utils`. The utils modules are: `checker_backends.py` and `utils.py`.

The file `base.py` contains the `BaseGrader` which the central class around this PR. It contains the basic logic of the grader, and should be inherited from to implement new graders. More about this in (About `BaseGrader`).

Currently there are two implementations at `runexe.py` and `simple.py`. `RunexeGrader` should works in exactly the same way as it works before this PR. `Simple` is a naive implementation of a grader without sandboxing (this is still WIP).

The `__init__.py` expose all the graders through the dictionary `GRADERS` and it can be selected with the new configuration added to `settings.ini` (`grader.GRADER_ID`). Current options are `runexe`/`simple`.

About `BaseGrader`:

Let's start with the docstring of `BaseGrader`:

    Cycle of the grader is as follows:

    1. First all arguments are gathered at `grade_submission`:

    2. Then the submission is compiled (only the submission, not extra files) with:
        ```
        grader.compile(arguments)
        ```

    3. The grader starts grading this submission.
        ```
        grader.run(arguments)
        ```

        This in itself is subdivided in two stages:

        3.1 Prepare extra sources.
            ```
            grader.prepare(arguments)
            ```

        3.2 Run submission against all testcases.
            ```
            grader.run_all()
            ```

            Run all will call `run_once(input_file, answer_file)`.
            - input_file: $PROBLEM/inputs/1.in
            - output_file: $PROBLEM/outputs/1.out

What is important about `BaseGrader` is that the function `run_once` is not implemented, and should be implemented by classes deriving it. It is expected that complex graders will need to override `prepare`, (and maybe `run_all`) but usually calling the superclass, since a lot of important variables are set there.

The following will describe how the full grading process works. When calling `grade_submission` `GraderArguments` will be built from the `submission` data, and it will afterwards call `grader.compile` and `grader.run`. If `grader.compile` returns not None, then it return `GraderResult` and it is because compilation failed somehow, so execution should stop at this moment, otherwise, we proceed to run this submission.

`grader.run` is expected to return `GraderResult` with the outcome of the execution, and contains all relevant information that should be added to the submission before completing, such as time-spent, memory-used, judgement-details, and most importantly verdict.

On the compilation stage several local variable to the grader are added, most notably `self.submission_folder` (which were all new data required for this submission should be created). Notice how we are using `pathlib` to build path instead of `os.path.join`.

On the `prepare` stage we create several local variables that can and should be used in `run_all` and `run_once`:
```
- self.arguments = arguments    # All arguments
- self.judgement_details = ''   # String builder, after running all testcase if solution is AC, this will be added as details
- self.checker_command = ...    # Command to execute checker.
- self.problem_folder = ...     # Folder where all testcases are located
- self.input_folder = ...       # Folder with input files.
- self.output_folder = ...      # Folder with output files.
- self.input_files = ...        # Sorted list of input files.
- self.output_files = ...       # Sorted list of output files.
```

On the `run_all` stage we iterate through every testcase and call `run_once` once per testcase. `run_once` should return `GraderPartialResult` with verdict from this single run, time spent and memory consumed. This partial results are aggregated and ultimately returned as `GraderResult` by `run_all`.