[problem] add a copy button to sample test cases
This was one of the top requests from contestants: the
ability to easily copy test cases from the problem page. In
this PR, I'm adding a simple copy button to each sample
input and output. When clicked, the content is copied to
the clipboard, and the button text changes to show a
checkmark. After 5 seconds, the checkmark disappears. I'm
not a frontend developer, so I'm not sure if this is the
best approach. :)