truncate very long samples
We're adding a problem to the qualifier with a
long sample test.

One option is to include a link for contestants to
download the test, but that adds extra work for
judges.

Another option is truncating and scrolling, but
users might not realize the sample is longer.

This PR truncates samples to 20 lines and shows a
warning indicating truncation. Users can click the
copy button to get the full content.