truncate very long samples (#176)
We're adding a problem to the qualifier with a
long sample test.

One option is to include a link for contestants to
download the test, but that adds extra work for
judges.

Another option is truncating and scrolling, but
users might not realize the sample is longer.

This PR truncates samples to 20 lines and shows
a ~warning indicating truncation. Users can click
the copy button to get the full content.

<img width="692" height="710" alt="image"
src="https://github.com/user-attachments/assets/86125783-efca-45be-b844-5910a1702ff7"
/>